Fix Anaconda environment configuration for tests
In order to fix scitt-community/scitt-api-emulator#38, we need to hard
pin the version of Werkzeug used by Flask to avoid the changes after
to how functions are organized in the library's modules.

See PR Sanster/lama-cleaner#390 with a related solution (not specific to
Anaconda approach, but informative).

See [StackOverflow](https://stackoverflow.com/a/77217971) post for
details.